requirements.txt
streamlit
st_pages
pickle
dynamicTreeCut
seaborn
PIL
matplotlib.pyplot
pandas
base64